refactor: move mindmap types to separate file
