+++ mindmapTypes.ts
@@ -1,0 +1,15 @@
+export interface NodeData {
+  id: string
+  x: number
+  y: number
+  label?: string
+  description?: string
+  parentId?: string | null
+  todoId?: string | null
+}
+
+export interface EdgeData {
+  id: string
+  from: string
+  to: string
+}
